Remove more code
